lab3/sleep/sleep/logic/requirements.txt
python-dotenv
flask_sqlalchemy==3.1.1
flask==3.1.0
pika==1.3.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.36
